Created SetupUserMixin
