+++ hackathon_site/event/test_api.py
@@ -2,28 +2,19 @@ from django.contrib.auth.models import Group
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
+from hackathon_site.tests import SetupUserMixin
 
 from event.models import Profile, User
 
 
-class CurrentUserTestCase(APITestCase):
+class CurrentUserTestCase(SetupUserMixin, APITestCase):
     def setUp(self):
-        self.password = "foobar123"
-        self.user = User.objects.create_user(
-            username="foo@bar.com",
-            password="foobar123",
-            first_name="Test",
-            last_name="Bar",
-            email="foo@bar.com",
-        )
+        super().setUp()
         self.group = Group.objects.create(name="Test Users")
         self.user.groups.add(self.group)
         self.profile = Profile.objects.create(user=self.user, status="Accepted")
 
         self.view = reverse("api:event:current-user")
-
-    def _login(self):
-        self.client.login(username=self.user.username, password=self.password)
 
     def test_user_not_logged_in(self):
         response = self.client.get(self.view)
